refactor(desktop): split web search provider adapters
